feat: extract vulkan display and resource modules
