+++ src/gpu/vulkan/mod.rs
@@ -3,12 +3,6 @@ use crate::utils::{
     Handle, Pool,
 };
 use ash::*;
-#[cfg(feature = "dashi-minifb")]
-use minifb;
-#[cfg(feature = "dashi-winit")]
-use winit;
-#[cfg(feature = "dashi-openxr")]
-use openxr as xr;
 pub use error::*;
 use std::{
     collections::HashMap,
@@ -38,6 +32,13 @@ pub mod openxr_window;
 mod conversions;
 pub use conversions::*;
 
+pub mod memory;
+pub use memory::*;
+pub mod image;
+pub use image::*;
+pub mod display;
+pub use display::*;
+
 /// Names of debugging layers that should be enabled when validation is requested.
 /// Only includes the standard Vulkan validation layer to avoid enabling any extra layers.
 pub const DEBUG_LAYER_NAMES: [*const c_char; 1] = [
@@ -58,155 +59,6 @@ unsafe extern "system" fn vulkan_debug_callback(
         message.to_string_lossy()
     );
     vk::FALSE
-}
-
-#[derive(Debug)]
-pub struct Buffer {
-    buf: vk::Buffer,
-    alloc: vk_mem::Allocation,
-    offset: u32,
-    size: u32,
-    suballocated: bool,
-}
-
-impl Clone for Buffer {
-    fn clone(&self) -> Self {
-        Self {
-            buf: self.buf.clone(),
-            alloc: unsafe { std::mem::transmute_copy(&self.alloc) },
-            offset: self.offset.clone(),
-            size: self.size.clone(),
-            suballocated: self.suballocated.clone(),
-        }
-    }
-}
-impl Handle<Buffer> {
-    /// Creates a [`DynamicBuffer`] view into this buffer at `byte_offset`.
-    ///
-    /// The caller must ensure the offset is within the buffer's bounds
-    /// and correctly aligned for the data that will be written.
-    pub fn to_unmapped_dynamic(&self, byte_offset: u32) -> DynamicBuffer {
-        return DynamicBuffer {
-            handle: self.clone(),
-            alloc: offset_allocator::Allocation {
-                offset: byte_offset,
-                metadata: 0,
-            },
-            ptr: std::ptr::null_mut(),
-            size: 0,
-        };
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct DynamicBuffer {
-    pub(crate) handle: Handle<Buffer>,
-    pub(crate) alloc: offset_allocator::Allocation,
-    pub(crate) ptr: *mut u8,
-    pub(crate) size: u16,
-}
-
-impl Default for DynamicBuffer {
-    fn default() -> Self {
-        Self {
-            handle: Default::default(),
-            alloc: offset_allocator::Allocation {
-                offset: 0,
-                metadata: 0,
-            },
-            ptr: std::ptr::null_mut(),
-            size: Default::default(),
-        }
-    }
-}
-impl DynamicBuffer {
-    /// Returns the handle to the underlying [`Buffer`].
-    pub fn handle(&self) -> Handle<Buffer> {
-        self.handle
-    }
-
-    /// Returns the byte offset into the underlying buffer.
-    pub fn offset(&self) -> u32 {
-        self.alloc.offset
-    }
-
-    /// Provides a mutable slice over the mapped memory.
-    ///
-    /// The slice is valid only while the buffer remains mapped and the
-    /// allocator that created it has not been reset or bumped.
-    pub fn slice<T>(&mut self) -> &mut [T] {
-        let typed_map: *mut T = unsafe { std::mem::transmute(self.ptr) };
-        return unsafe {
-            std::slice::from_raw_parts_mut(typed_map, self.size as usize / std::mem::size_of::<T>())
-        };
-    }
-}
-
-#[derive(Clone)]
-pub struct DynamicAllocator {
-    allocator: offset_allocator::Allocator,
-    ptr: *mut u8,
-    min_alloc_size: u32,
-    pool: Handle<Buffer>,
-}
-
-impl Default for DynamicAllocator {
-    fn default() -> Self {
-        Self {
-            allocator: Default::default(),
-            ptr: std::ptr::null_mut(),
-            min_alloc_size: Default::default(),
-            pool: Default::default(),
-        }
-    }
-}
-
-impl DynamicAllocator {
-    /// Resets the allocator, invalidating all previously bumped buffers.
-    ///
-    /// # Safety
-    /// Only call this when no GPU operations are reading from the
-    /// allocations produced by this allocator.
-    pub fn reset(&mut self) {
-        self.allocator.reset();
-    }
-
-    /// Allocates a new [`DynamicBuffer`] of `min_alloc_size` bytes.
-    ///
-    /// # Safety
-    /// The caller must ensure that no pending GPU reads exist for
-    /// buffers previously returned by this allocator before calling this
-    /// method.
-    pub fn bump(&mut self) -> Option<DynamicBuffer> {
-        let alloc = self.allocator.allocate(self.min_alloc_size)?;
-        return Some(DynamicBuffer {
-            handle: self.pool,
-            alloc,
-            ptr: unsafe { self.ptr.offset(alloc.offset as isize) },
-            size: self.min_alloc_size as u16,
-        });
-    }
-}
-#[derive(Debug)]
-pub struct Image {
-    img: vk::Image,
-    alloc: vk_mem::Allocation,
-    dim: [u32; 3],
-    format: Format,
-    layouts: Vec<vk::ImageLayout>,
-    sub_layers: vk::ImageSubresourceLayers,
-}
-
-#[derive(Debug)]
-pub struct Sampler {
-    sampler: vk::Sampler,
-}
-
-#[derive(Debug)]
-pub struct ImageView {
-    img: Handle<Image>,
-    range: vk::ImageSubresourceRange,
-    view: vk::ImageView,
 }
 
 #[derive(Clone, Default)]
@@ -259,112 +111,6 @@ pub struct IndexedBindGroup {
     set_id: u32,
 }
 
-#[allow(dead_code)]
-pub struct Display {
-    #[cfg(all(feature = "dashi-sdl2", not(feature = "dashi-openxr")))]
-    window: std::cell::Cell<sdl2::video::Window>,
-    #[cfg(all(feature = "dashi-minifb", not(feature = "dashi-openxr")))]
-    window: minifb::Window,
-    #[cfg(all(feature = "dashi-winit", not(feature = "dashi-openxr")))]
-    window: winit::window::Window,
-    #[cfg(all(feature = "dashi-winit", not(feature = "dashi-openxr")))]
-    event_loop: winit::event_loop::EventLoop<()>,
-    #[cfg(not(feature = "dashi-openxr"))]
-    swapchain: ash::vk::SwapchainKHR,
-    #[cfg(not(feature = "dashi-openxr"))]
-    surface: ash::vk::SurfaceKHR,
-    #[cfg(not(feature = "dashi-openxr"))]
-    images: Vec<Handle<Image>>,
-    #[cfg(not(feature = "dashi-openxr"))]
-    views: Vec<Handle<ImageView>>,
-    #[cfg(not(feature = "dashi-openxr"))]
-    loader: ash::extensions::khr::Surface,
-    #[cfg(not(feature = "dashi-openxr"))]
-    sc_loader: ash::extensions::khr::Swapchain,
-    #[cfg(not(feature = "dashi-openxr"))]
-    semaphores: Vec<Handle<Semaphore>>,
-    #[cfg(not(feature = "dashi-openxr"))]
-    fences: Vec<Handle<Fence>>,
-    #[cfg(not(feature = "dashi-openxr"))]
-    frame_idx: u32,
-
-    #[cfg(feature = "dashi-openxr")]
-    xr_instance: xr::Instance,
-    #[cfg(feature = "dashi-openxr")]
-    xr_session: xr::Session<xr::Vulkan>,
-    #[cfg(feature = "dashi-openxr")]
-    xr_waiter: xr::FrameWaiter,
-    #[cfg(feature = "dashi-openxr")]
-    xr_stream: xr::FrameStream<xr::Vulkan>,
-    #[cfg(feature = "dashi-openxr")]
-    xr_swapchain: xr::Swapchain<xr::Vulkan>,
-    #[cfg(feature = "dashi-openxr")]
-    xr_images: Vec<vk::Image>,
-    #[cfg(feature = "dashi-openxr")]
-    xr_view_config: Vec<xr::ViewConfigurationView>,
-    #[cfg(feature = "dashi-openxr")]
-    xr_input: openxr_window::XrInput,
-}
-
-impl Display {
-    #[cfg(all(feature = "dashi-minifb", not(feature = "dashi-openxr")))]
-    pub fn minifb_window(&mut self) -> &mut minifb::Window {
-        &mut self.window
-    }
-    #[cfg(all(feature = "dashi-winit", not(feature = "dashi-openxr")))]
-    pub fn winit_window(&self) -> &winit::window::Window {
-        &self.window
-    }
-    #[cfg(all(feature = "dashi-winit", not(feature = "dashi-openxr")))]
-    pub fn winit_event_loop(&mut self) -> &mut winit::event_loop::EventLoop<()> {
-        &mut self.event_loop
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn xr_instance(&self) -> &xr::Instance {
-        &self.xr_instance
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn xr_session(&self) -> &xr::Session<xr::Vulkan> {
-        &self.xr_session
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn xr_frame_waiter(&mut self) -> &mut xr::FrameWaiter {
-        &mut self.xr_waiter
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn xr_frame_stream(&mut self) -> &mut xr::FrameStream<xr::Vulkan> {
-        &mut self.xr_stream
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn xr_swapchain(&self) -> &xr::Swapchain<xr::Vulkan> {
-        &self.xr_swapchain
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn xr_swapchain_images(&self) -> &[vk::Image] {
-        &self.xr_images
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn xr_view_configuration(&self) -> &[xr::ViewConfigurationView] {
-        &self.xr_view_config
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn poll_inputs(&mut self) -> xr::Result<openxr_window::XrInputState> {
-        self.xr_input.poll_inputs()
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    pub fn xr_input(&self) -> &openxr_window::XrInput {
-        &self.xr_input
-    }
-}
 
 
 #[derive(Clone)]
@@ -3370,386 +3116,6 @@ impl Context {
                 .create_framebuffer(&framebuffer_info, None)
                 .expect("Failed to create framebuffer")
         });
-    }
-
-    #[cfg(not(feature = "dashi-openxr"))]
-    /// Destroys a windowing display and its swapchain.
-    ///
-    /// # Prerequisites
-    /// - Ensure the GPU has finished using all swapchain images.
-    /// - Image views should be destroyed before their images and swapchain.
-    /// - The context must still be alive.
-    pub fn destroy_display(&mut self, dsp: Display) {
-        for img in &dsp.images {
-            self.images.release(*img);
-        }
-
-        for img in &dsp.views {
-            self.destroy_image_view(*img);
-        }
-        unsafe { dsp.sc_loader.destroy_swapchain(dsp.swapchain, None) };
-        unsafe { dsp.loader.destroy_surface(dsp.surface, None) };
-        for sem in dsp.semaphores {
-            self.destroy_semaphore(sem);
-        }
-
-        for fence in &dsp.fences {
-            self.destroy_fence(fence.clone());
-        }
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    /// Destroys an OpenXR display.
-    ///
-    /// # Prerequisites
-    /// - Ensure the GPU has finished using any swapchain images provided by OpenXR.
-    /// - The context must still be alive.
-    ///
-    /// OpenXR resources are cleaned up by their `Drop` implementations.
-    pub fn destroy_display(&mut self, _dsp: Display) {
-        // OpenXR resources are cleaned up by Drop implementations
-    }
-
-    #[cfg(all(feature = "dashi-sdl2", not(feature = "dashi-openxr")))]
-    fn make_window(
-        &mut self,
-        info: &WindowInfo,
-    ) -> (std::cell::Cell<sdl2::video::Window>, vk::SurfaceKHR) {
-        let mut window = std::cell::Cell::new(
-            self.sdl_video
-                .as_ref()
-                .unwrap()
-                .window(&info.title, info.size[0], info.size[1])
-                .vulkan()
-                .build()
-                .expect("Unable to create SDL2 Window!"),
-        );
-
-        let surface = window
-            .get_mut()
-            .vulkan_create_surface(vk::Handle::as_raw(self.instance.handle()) as usize)
-            .expect("Unable to create vulkan surface!");
-
-        (window, vk::Handle::from_raw(surface))
-    }
-
-    #[cfg(all(feature = "dashi-minifb", not(feature = "dashi-openxr")))]
-    fn make_window(
-        &mut self,
-        info: &WindowInfo,
-    ) -> (minifb::Window, vk::SurfaceKHR) {
-        minifb_window::create_window(&self.entry, &self.instance, info).unwrap()
-    }
-
-    #[cfg(all(feature = "dashi-winit", not(feature = "dashi-openxr")))]
-    fn make_window(
-        &mut self,
-        info: &WindowInfo,
-    ) -> (winit::event_loop::EventLoop<()>, winit::window::Window, vk::SurfaceKHR) {
-        winit_window::create_window(&self.entry, &self.instance, info).unwrap()
-    }
-
-    #[cfg(not(feature = "dashi-openxr"))]
-    /// Creates a display and associated swapchain for presenting images.
-    ///
-    /// # Prerequisites
-    /// - Correct attachment formats.
-    /// - Matching pipeline layouts.
-    /// - Swapchain acquisition order is respected.
-    /// - XR session state is valid.
-    /// - Synchronization primitives are handled during presentation.
-    pub fn make_display(&mut self, info: &DisplayInfo) -> Result<Display, GPUError> {
-        if self.headless {
-            return Err(GPUError::HeadlessDisplayNotSupported);
-        }
-        #[cfg(feature = "dashi-winit")]
-        let (event_loop, window, surface) = self.make_window(&info.window);
-        #[cfg(not(feature = "dashi-winit"))]
-        let (window, surface) = self.make_window(&info.window);
-
-        let loader = ash::extensions::khr::Surface::new(&self.entry, &self.instance);
-        let capabilities =
-            unsafe { loader.get_physical_device_surface_capabilities(self.pdevice, surface)? };
-        let _formats =
-            unsafe { loader.get_physical_device_surface_formats(self.pdevice, surface)? };
-        let _present_modes =
-            unsafe { loader.get_physical_device_surface_present_modes(self.pdevice, surface)? };
-
-        // Choose extent
-        let size = info.window.size;
-        let mut chosen_extent = vk::Extent2D {
-            width: size[0],
-            height: size[1],
-        };
-        if capabilities.current_extent.width != std::u32::MAX {
-            chosen_extent = capabilities.current_extent.clone();
-        } else {
-            chosen_extent.width = std::cmp::max(
-                capabilities.min_image_extent.width,
-                std::cmp::min(capabilities.max_image_extent.width, chosen_extent.width),
-            );
-            chosen_extent.height = std::cmp::max(
-                capabilities.min_image_extent.height,
-                std::cmp::min(capabilities.max_image_extent.height, chosen_extent.height),
-            );
-        }
-
-        // Select a present mode.
-        let present_mode = if info.vsync {
-            vk::PresentModeKHR::FIFO
-        } else {
-            vk::PresentModeKHR::IMMEDIATE
-        };
-
-        let wanted_format = vk::Format::B8G8R8A8_SRGB;
-        let wanted_color_space = vk::ColorSpaceKHR::SRGB_NONLINEAR;
-
-        let image_usage = vk::ImageUsageFlags::TRANSFER_DST
-            | vk::ImageUsageFlags::TRANSFER_SRC
-            | vk::ImageUsageFlags::COLOR_ATTACHMENT;
-
-        let num_framebuffers = match info.buffering {
-            WindowBuffering::Double => 2,
-            WindowBuffering::Triple => 3,
-        };
-        let swap_loader = ash::extensions::khr::Swapchain::new(&self.instance, &self.device);
-        let swapchain = unsafe {
-            swap_loader.create_swapchain(
-                &vk::SwapchainCreateInfoKHR::builder()
-                    .surface(surface)
-                    .present_mode(present_mode)
-                    .image_format(wanted_format)
-                    .image_array_layers(1)
-                    .image_color_space(wanted_color_space)
-                    .image_sharing_mode(vk::SharingMode::EXCLUSIVE)
-                    .image_extent(chosen_extent)
-                    .image_usage(image_usage)
-                    .min_image_count(std::cmp::max(
-                        num_framebuffers,
-                        capabilities.min_image_count,
-                    ))
-                    .pre_transform(capabilities.current_transform)
-                    .composite_alpha(vk::CompositeAlphaFlagsKHR::OPAQUE)
-                    .build(),
-                Default::default(),
-            )?
-        };
-
-        // Now, we need to make the images!
-        let images = unsafe { swap_loader.get_swapchain_images(swapchain)? };
-        let mut handles: Vec<Handle<Image>> = Vec::with_capacity(images.len() as usize);
-        let mut view_handles: Vec<Handle<ImageView>> = Vec::with_capacity(images.len() as usize);
-        for img in images {
-            match self.images.insert(Image {
-                img,
-                alloc: unsafe { std::mem::MaybeUninit::zeroed().assume_init() },
-                layouts: vec![vk::ImageLayout::UNDEFINED],
-                sub_layers: vk::ImageSubresourceLayers::builder()
-                    .layer_count(1)
-                    .mip_level(0)
-                    .base_array_layer(0)
-                    .aspect_mask(vk::ImageAspectFlags::COLOR)
-                    .build(),
-                dim: [chosen_extent.width, chosen_extent.height, 1],
-                format: vk_to_lib_image_format(wanted_format),
-            }) {
-                Some(handle) => {
-                    self.oneshot_transition_image_noview(handle, vk::ImageLayout::PRESENT_SRC_KHR);
-                    let h = self.make_image_view(&ImageViewInfo {
-                        debug_name: &info.window.title,
-                        img: handle,
-                        layer: 0,
-                        mip_level: 0,
-                        ..Default::default()
-                    })?;
-
-                    view_handles.push(h);
-                    handles.push(handle)
-                }
-                None => todo!(),
-            };
-        }
-
-        let sems = self.make_semaphores(handles.len()).unwrap();
-        let mut fences = Vec::with_capacity(handles.len() as usize);
-        for _idx in 0..handles.len() {
-            fences.push(self.make_fence()?);
-        }
-        self.init_gpu_timers(handles.len())?;
-        return Ok(Display {
-            window,
-            #[cfg(feature = "dashi-winit")]
-            event_loop,
-            swapchain,
-            surface,
-            images: handles,
-            loader,
-            sc_loader: swap_loader,
-            frame_idx: 0,
-            semaphores: sems,
-            fences,
-            views: view_handles,
-        });
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    /// Creates an XR display and session for immersive rendering.
-    ///
-    /// # Prerequisites
-    /// - Correct attachment formats.
-    /// - Matching pipeline layouts.
-    /// - Swapchain acquisition order is respected.
-    /// - XR session state is valid.
-    /// - Synchronization primitives are handled during presentation.
-    pub fn make_xr_display(&mut self, _info: &XrDisplayInfo) -> Result<Display, GPUError> {
-        let (xr_instance, session, waiter, stream, swapchain, images, views) =
-            openxr_window::create_xr_session(
-                &self.instance,
-                self.pdevice,
-                &self.device,
-                self.gfx_queue.family,
-            )
-            .map_err(|_| GPUError::LibraryError())?;
-
-        let xr_input = openxr_window::XrInput::new(&xr_instance, &session)
-            .map_err(|_| GPUError::LibraryError())?;
-
-        Ok(Display {
-            xr_instance,
-            xr_session: session,
-            xr_waiter: waiter,
-            xr_stream: stream,
-            xr_swapchain: swapchain,
-            xr_images: images,
-            xr_view_config: views,
-            xr_input,
-        })
-    }
-
-    #[cfg(not(feature = "dashi-openxr"))]
-    /// Acquires the next image from a windowed swapchain.
-    ///
-    /// # Prerequisites
-    /// - Correct attachment formats.
-    /// - Matching pipeline layouts.
-    /// - Swapchain acquisition order is respected.
-    /// - XR session state is valid.
-    /// - Synchronization primitives are handled during presentation.
-    pub fn acquire_new_image(
-        &mut self,
-        dsp: &mut Display,
-    ) -> Result<(Handle<ImageView>, Handle<Semaphore>, u32, bool), GPUError> {
-        let signal_sem_handle = dsp.semaphores[dsp.frame_idx as usize].clone();
-        let fence = dsp.fences[dsp.frame_idx as usize];
-
-        self.wait(fence.clone())?;
-
-        let signal_sem = self.semaphores.get_ref(signal_sem_handle).unwrap();
-        let res = unsafe {
-            dsp.sc_loader.acquire_next_image2(
-                &vk::AcquireNextImageInfoKHR::builder()
-                    .swapchain(dsp.swapchain)
-                    .semaphore(signal_sem.raw)
-                    .fence(
-                        self.fences
-                            .get_ref(dsp.fences[dsp.frame_idx as usize])
-                            .unwrap()
-                            .raw,
-                    )
-                    .timeout(std::u64::MAX)
-                    .device_mask(0x1)
-                    .build(),
-            )
-        }?;
-
-        dsp.frame_idx = res.0;
-        return Ok((dsp.views[res.0 as usize], signal_sem_handle, res.0, res.1));
-    }
-
-    #[cfg(not(feature = "dashi-openxr"))]
-    /// Presents the current swapchain image to the display.
-    ///
-    /// # Prerequisites
-    /// - Correct attachment formats.
-    /// - Matching pipeline layouts.
-    /// - Swapchain acquisition order is respected.
-    /// - XR session state is valid.
-    /// - Synchronization primitives are handled during presentation.
-    pub fn present_display(
-        &mut self,
-        dsp: &Display,
-        wait_sems: &[Handle<Semaphore>],
-    ) -> Result<(), GPUError> {
-        let mut raw_wait_sems: Vec<vk::Semaphore> = Vec::with_capacity(32);
-        for sem in wait_sems {
-            raw_wait_sems.push(self.semaphores.get_ref(sem.clone()).unwrap().raw);
-        }
-
-        unsafe {
-            dsp.sc_loader.queue_present(
-                self.gfx_queue.queue,
-                &vk::PresentInfoKHR::builder()
-                    .image_indices(&[dsp.frame_idx])
-                    .swapchains(&[dsp.swapchain])
-                    .wait_semaphores(&raw_wait_sems)
-                    .build(),
-            )
-        }?;
-
-        return Ok(());
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    /// Acquires the next image from the XR swapchain.
-    ///
-    /// # Prerequisites
-    /// - Correct attachment formats.
-    /// - Matching pipeline layouts.
-    /// - Swapchain acquisition order is respected.
-    /// - XR session state is valid.
-    /// - Synchronization primitives are handled during presentation.
-    pub fn acquire_xr_image(
-        &mut self,
-        dsp: &mut Display,
-    ) -> Result<(u32, xr::FrameState), GPUError> {
-        let state = dsp
-            .xr_waiter
-            .wait()
-            .map_err(|_| GPUError::LibraryError())?;
-        dsp.xr_stream
-            .begin()
-            .map_err(|_| GPUError::LibraryError())?;
-        let idx = dsp
-            .xr_swapchain
-            .acquire_image()
-            .map_err(|_| GPUError::LibraryError())?;
-        dsp.xr_swapchain
-            .wait_image(xr::Duration::INFINITE)
-            .map_err(|_| GPUError::LibraryError())?;
-        Ok((idx, state))
-    }
-
-    #[cfg(feature = "dashi-openxr")]
-    /// Presents the current frame to the XR display.
-    ///
-    /// # Prerequisites
-    /// - Correct attachment formats.
-    /// - Matching pipeline layouts.
-    /// - Swapchain acquisition order is respected.
-    /// - XR session state is valid.
-    /// - Synchronization primitives are handled during presentation.
-    pub fn present_xr_display(
-        &mut self,
-        dsp: &mut Display,
-        state: xr::FrameState,
-    ) -> Result<(), GPUError> {
-        dsp.xr_swapchain
-            .release_image()
-            .map_err(|_| GPUError::LibraryError())?;
-        dsp.xr_stream
-            .end(state.predicted_display_time, xr::EnvironmentBlendMode::OPAQUE, &[])
-            .map_err(|_| GPUError::LibraryError())?;
-        Ok(())
     }
 }
